+++ wemake_python_styleguide/types.py
@@ -273,3 +273,7 @@ class ConfigurationOptions(Protocol):
     @property
     def show_violation_links(self) -> bool:
         ...
+
+    @property
+    def disallow_union_type(self) -> bool:
+        ...